`usb_bus_ehci`: move register structs to a separate module
`main.rs` is going to become very unwieldly as is, so these can be separated
out.
